outsource torch object
