+++ ncopt/sqpgs.py
@@ -1,3 +1,6 @@
+"""
+author: Fabian Schaipp
+"""
 import numpy as np
 import cvxopt as cx
